requirements.txt
# Video Transcription Pipeline - Requirements
# ============================================
# Tested with Python 3.12+ and CUDA 12.4
#
# Installation order matters for some packages.
# See README.md for detailed installation instructions.

# =============================================================================
# Core Dependencies
# =============================================================================

# PyTorch with CUDA support (install first)
# Use: pip install torch torchaudio --index-url https://download.pytorch.org/whl/cu124
torch>=2.6.0
torchaudio>=2.6.0

# Transformers (Granite requires >=4.52.4)
transformers>=4.52.4

# Image processing
pillow>=11.0.0
imagehash>=4.3.0

# =============================================================================
# Gemini Scripts (process_video_gemini_*.py)
# =============================================================================

# Google Gen AI SDK for Gemini API
google-genai>=1.0.0

# =============================================================================
# ASR Models
# =============================================================================

# WhisperX (recommended for timestamps)
# Install via: pip install whisperx
# whisperx  # Not on PyPI, install separately

# OpenAI Whisper
openai-whisper>=20231117

# Granite Speech requirements
peft>=0.15.0
soundfile>=0.12.0

# Canary/Parakeet (NeMo toolkit)
# Install via: pip install "nemo_toolkit[asr] @ git+https://github.com/NVIDIA/NeMo.git"
# nemo_toolkit  # Install from git, not PyPI

# =============================================================================
# Local-Only Script (process_video_local_only.py)
# =============================================================================

# Qwen2.5-VL utilities
qwen-vl-utils>=0.0.8

# AutoAWQ for quantized models (Qwen2.5-14B-AWQ)
# NOTE: AutoAWQ is deprecated but still functional with torch 2.6.0
# Install via: pip install autoawq
# autoawq  # May have compatibility issues, install separately

# =============================================================================
# Pinned Versions (tested configuration)
# =============================================================================
# If you encounter issues, try these exact versions:
#
# torch==2.6.0+cu124
# torchaudio==2.6.0+cu124
# transformers==4.57.5
# google-genai==1.57.0
# openai-whisper==20250625
# pillow==12.0.0
# imagehash==4.3.2
# peft==0.18.1
# soundfile==0.13.1
# qwen-vl-utils==0.0.8
